// src/api/dto/utils/request/HttpMethod.ts
/**
 * http请求方法
 *
 * @author LCTR
 * @date 2023-03-09
 */
export const enum HttpMethod {
  GET = 'GET',
  DELETE = 'DELETE',
  HEAD = 'HEAD',
  OPTIONS = 'OPTIONS',
  POST = 'POST',
  PUT = 'PUT',
  PATCH = 'PATCH',
  PURGE = 'PURGE',
  LINK = 'LINK',
  UNLINK = 'UNLINK',
}